[CALCITE-3333] Add pluggable frame size limits
Add an ability to plug in custom limiting of ResultSet frame
sizes by overriding a method of JdbcMeta.

The intention is to allow limiting the size of a ResultSet
frame based on the amount of time taken to create it, in addition
to limiting based on the number of rows in the frame.