[CALCITE-4196] Consume all data from client before replying with HTTP/401
SPNEGO's handshake involves sending an HTTP/401 to "challenge" the
client to reply with authentication data. If the client is sending
a significant amount of data in the original request, the client
will still be writing this data when the server replies. This causes
the client to receive a TCP Reset when it continues to write data, and
ultimately manifests in a "Broken Pipe" runtime exception.

The fix is to simply consume all data the client wrote prior to
responding with the HTTP/401.

Closes #127

Signed-off-by: Kevin Risden <krisden@apache.org>