[CALCITE-4152] Upgrade Avatica to use the configurable SPNEGO Jetty implementation
Jetty has deprecated the previously-used version of SPNEGO login code.
This change requires a few other changes to adopt:

1. Removal of automatic server login via JAAS (Jetty removed this and
    expects explicit logins for the server).
2. Separation of Authentication and Authorization (we're required to
    use a LoginService for authz to use the new SPNEGO authentication).

For the benefit of making this change, we automatically inherit the
Jetty Session logic which can skip SPNEGO authentication for the 2nd
to Nth call to Avatica. For a "workload" which previously took N HTTP calls
to Avatica to perform, this can now be done in (N/2)+1 HTTP calls
which, for average Avatica calls, results in a nearly 2x speed-up.

Jetty Sessions will cause a JSESSIONID cookie to be sent back on the
successful SPNEGO authentication handshake. As long as the client
resubmits this cookie for subsequent requests, the identity of the
client is kept intact.

To test this more easily, this change also includes updates to the
Avatica StandaloneServer, which more easily enables setup of Avatica
against any database (e.g. hsqldb with the SCOTT dataset).