[CALCITE-5009] Transparent JDBC connection re-creation may lead to data loss
CALCITE-903 has introduced a transaprent reconnection feature, which will open
a new server-side connection in case it is expired from the server side connection
cache.

While this is convinient for most read-only analytical workload, this
can cause a number a problems, including data loss for transactional connections.

This patch disables the transparent reconnect feature by default, and adds the
transparent_reconnection property, which re-enables it when set to true.