[CALCITE-5338] In DateTimeUtils, deprecate floorMod and floorDiv, and use equivalents in java.lang.Math
In Avatica's DateTimeUtils, replace floorDiv and floorMod
with equivalents in java.lang.Math.

The JDK equivalents were introduced in Java 8, probably have
superior performance (due to intrinsics), and we can use
them now that Avatica is JDK 8 and above.

java.lang.Math.floorMod(long, int), which was introduced in
JDK 9, is still off-limits.